refactor: favorite 패키지를 bookmark 패키지로 이름 변경 및 관련 코드 수정
