+++ android/app/src/main/java/com/on/turip/domain/bookmark/repository/BookmarkRepository.kt
@@ -1,7 +1,7 @@
-package com.on.turip.domain.favorite.repository
+package com.on.turip.domain.bookmark.repository
 
 import com.on.turip.core.result.TuripResult
-import com.on.turip.domain.favorite.PagedBookmarkContents
+import com.on.turip.domain.bookmark.PagedBookmarkContents
 
 interface BookmarkRepository {
     suspend fun createBookmark(contentId: Long): TuripResult<Unit>
